BUG FIX: Edit profile page not saving States correctly.
BUG FIX: When updating a user’s profile, state field would not save or
update values.

This was caused by JS replacing the state fields as a dropdown and the
ID value (by PMPro default) was different to each other.

Checkout page ID for state was ‘#bstate’ whereas loading the state
fields in the edit user profile was ‘#pmpro_bstate’.